Extract base class
