+++ src/Roslyn.Diagnostics.Analyzers/CSharp/AnalyzerReleases.Unshipped.md
@@ -2,4 +2,3 @@
 ### New Rules
 Rule ID | Category | Severity | Notes
 --------|----------|----------|-------
-RS0103 | RoslynDiagnosticsMaintainability | Warning | CSharpBlankLinesBetweenStatementsDiagnosticAnalyzer
